+++ Controller Classes/PatchTableDataSource.h
@@ -14,8 +14,8 @@
 
 - (void)setPatchArray:(NSMutableArray*)newPatchArray;
 
-- (int)numberOfRowsInTableView:(NSTableView*)tableView;
-- (id)tableView:(NSTableView*)tableView objectValueForTableColumn:(NSTableColumn*)column row:(int)rowIndex;
+- (NSUInteger)numberOfRowsInTableView:(NSTableView*)tableView;
+- (id)tableView:(NSTableView*)tableView objectValueForTableColumn:(NSTableColumn*)column row:(NSUInteger)rowIndex;
 - (void)tableView:(NSTableView*)tableView setObjectValue:(id)value forTableColumn:(NSTableColumn*)column row:(int)rowIndex;
 - (void)deleteSelection:(NSTableView*)tableView;
