+++ Controller Classes/PatchbayDocument.h
@@ -90,9 +90,9 @@
 - (void)selectedPatchChanged:(NSNotification*)notification;
 - (IBAction)addPatchButtonPressed:(id)sender;
 
-- (void)addPatch:(Patch*)patch atIndex:(int)index;
-- (void)addPatchFromArchive:(NSData*)data atIndex:(int)index;
-- (void)removePatchAtIndex:(int)index;
+- (void)addPatch:(Patch*)patch atIndex:(NSUInteger)index;
+- (void)addPatchFromArchive:(NSData*)data atIndex:(NSUInteger)index;
+- (void)removePatchAtIndex:(NSUInteger)index;
 - (void)setIsEnabled:(BOOL)isEnabled forPatch:(Patch*)patch;
 
 - (NSData*)archivePatchForPasteBoard:(Patch*)patch;
